further main.cpp break down
